python_mqtt_bridge/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
altgraph==0.17.3
astroid==2.12.11
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
bandit==1.7.4
bidict==0.22.0
blinker==1.5
cadquery==1.2.5
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
commonmark==0.9.1
cx-Freeze==6.13.0
dearpygui==1.6.2
dill==0.3.5.1
docutils==0.19
dodgy==0.2.1
feedgenerator==2.0.0
flake8==4.0.1
flake8-polyfill==1.0.2
frozenlist==1.3.1
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.3.post0
idna==3.4
isort==5.10.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
macholib==1.16.2
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.6.1
modulegraph==0.19.3
mpmath==1.2.1
msgpack==1.0.4
multidict==6.0.2
mypy==0.982
mypy-extensions==0.4.3
npyscreen==4.10.5
numpy==1.23.1
opencv-python==4.5.4.58
packaging==21.3
paho-mqtt==1.6.1
pbr==5.10.0
pelican==4.8.0
pep8-naming==0.10.0
Pillow==9.2.0
platformdirs==2.5.2
prospector==1.7.7
pycodestyle==2.8.0
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.13.0
pyinstaller-hooks-contrib==2022.11
pylint==2.15.4
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pynvim==0.4.3
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-osc==1.8.0
python-socketio==5.7.2
pytz==2022.4
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.3.1
requests==2.28.1
requirements-detector==0.7
rich==12.6.0
setoptconf-tmp==0.3.1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
stevedore==4.0.1
sympy==1.11.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
Unidecode==1.3.6
urllib3==1.26.12
websocket-client==1.4.1
wrapt==1.14.1
yarl==1.8.1
